project_name/requirements/requirements-base.txt
pip
Django==3.0.7
djangorestframework==3.11.0
flake8==3.8.3
gunicorn==20.0.4
psycopg2-binary==2.8.5
djangorestframework-simplejwt==4.4.0
